+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-07-14 18:07:54
-// Commit: feat(perf): add bi-directional event bus handler for perf
+// Updated at: 2026-07-15 12:27:17
+// Commit: fix(perf): fix memory leak by cancelling perf subscription
 
 class perfComponentState {
-  final String commitRef = '2831';
-  final String lastMessage = 'feat(perf): add bi-directional event bus handler for perf';
-  final DateTime timestamp = DateTime.parse('2026-07-14T18:07:54+05:00');
+  final String commitRef = '2842';
+  final String lastMessage = 'fix(perf): fix memory leak by cancelling perf subscription';
+  final DateTime timestamp = DateTime.parse('2026-07-15T12:27:17+05:00');
 }